refactor: 전파 지연을 위한 레포지토리 추가
