+++ src/main/java/coupon/repository/WriterRepository.java
@@ -1,6 +1,8 @@
 package coupon.repository;
 
 import coupon.entity.CouponEntity;
+import coupon.entity.UserCouponEntity;
+import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -9,12 +11,18 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Component
 @RequiredArgsConstructor
-public class CouponWriterRepository {
+public class WriterRepository {
 
     private final CouponRepository couponRepository;
+    private final UserCouponRepository userCouponRepository;
 
     @Transactional(propagation = Propagation.REQUIRES_NEW)
-    public Optional<CouponEntity> readFromWriter(long id) {
+    public Optional<CouponEntity> findCouponById(long id) {
         return couponRepository.findById(id);
+    }
+
+    @Transactional(propagation = Propagation.REQUIRES_NEW)
+    public List<UserCouponEntity> findUserCouponsByUserId(long userId) {
+        return userCouponRepository.findAllByUserId(userId);
     }
 }
