+++ src/main/java/coupon/service/CouponService.java
@@ -4,7 +4,7 @@ import coupon.domain.coupon.Coupon;
 import coupon.entity.CouponEntity;
 import coupon.mapper.DomainEntityMapper;
 import coupon.mapper.EntityDomainMapper;
-import coupon.repository.CouponReaderRepository;
+import coupon.repository.ReaderRepository;
 import coupon.repository.CouponRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -15,7 +15,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class CouponService {
 
     private final CouponRepository couponRepository;
-    private final CouponReaderRepository couponReaderRepository;
+    private final ReaderRepository readerRepository;
 
     @Transactional
     public CouponEntity create(Coupon coupon) {
@@ -24,7 +24,7 @@ public class CouponService {
 
     @Transactional(readOnly = true)
     public Coupon getCoupon(Long id) {
-        CouponEntity couponEntity = couponReaderRepository.findById(id)
+        CouponEntity couponEntity = readerRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 쿠폰입니다."));
         return EntityDomainMapper.mapToCoupon(couponEntity);
     }
